consolidate responsive logic into Responsive Manager
